+++ .fallowrc.json
@@ -5,7 +5,7 @@
     "src/index.ts",
     "src/io.ts",
     "src/artifacts.ts",
-    "src/metro.ts",
+    "src/metro/metro.ts",
     "src/remote-config.ts",
     "src/install-source.ts",
     "src/android-adb.ts",
