+++ src/metro/metro.ts
@@ -1,4 +1,4 @@
-import type { SessionRuntimeHints } from './contracts.ts';
+import type { SessionRuntimeHints } from '../contracts.ts';
 import {
   buildMetroRuntimeHints,
   prepareMetroRuntime,
@@ -8,10 +8,13 @@ import {
   type ReloadMetroResult,
 } from './client-metro.ts';
 import { ensureMetroCompanion, stopMetroCompanion } from './client-metro-companion.ts';
-import type { MetroBridgeScope } from './client-companion-tunnel-contract.ts';
-import { resolveRuntimeTransportHints } from './utils/runtime-transport.ts';
-export { buildBundleUrl, normalizeBaseUrl } from './utils/url.ts';
-export type { CompanionTunnelScope, MetroBridgeScope } from './client-companion-tunnel-contract.ts';
+import type { MetroBridgeScope } from '../client-companion-tunnel-contract.ts';
+import { resolveRuntimeTransportHints } from '../utils/runtime-transport.ts';
+export { buildBundleUrl, normalizeBaseUrl } from '../utils/url.ts';
+export type {
+  CompanionTunnelScope,
+  MetroBridgeScope,
+} from '../client-companion-tunnel-contract.ts';
 
 export type {
   MetroBridgeDescriptor,
